+++ src/components/Form/inputField/index.tsx
@@ -1,0 +1,38 @@
+import React from 'react';
+import * as S from './styles';
+
+interface InputFieldProps {
+  type: string;
+  placeholder: string;
+  value: string;
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  onBlur: (e: React.FocusEvent<HTMLInputElement>) => void;
+  hasError?: boolean;
+  error?: string;
+}
+
+const InputField: React.FC<InputFieldProps> = ({
+  type,
+  placeholder,
+  value,
+  onChange,
+  onBlur,
+  hasError,
+  error
+}) => {
+  return (
+    <S.InputContainer>
+      <S.Input
+        type={type}
+        placeholder={placeholder}
+        value={value}
+        onChange={onChange}
+        onBlur={onBlur}
+        hasError={hasError}
+      />
+      {error && <S.ErrorMessage>{error}</S.ErrorMessage>}
+    </S.InputContainer>
+  );
+};
+
+export default InputField;
