+++ src/pages/Home/Page.tsx
@@ -1,7 +1,7 @@
-import FriendSelector from '../../components/FriendSelector';
-import CategorySection from '../../components/CategorySection';
-import MessageSection from '../../components/MessageSection';
-import RankingSection from '../../components/RankingSection';
+import FriendSelector from './components/FriendSelector';
+import CategorySection from './components/CategorySection';
+import MessageSection from './components/MessageSection';
+import RankingSection from './components/RankingSection';
 
 const Home = () => (
   <>
