+++ src/pages/Home/components/CategorySection/styles.ts
