+++ src/pages/Home/components/FriendSelector/index.tsx
