+++ src/pages/Home/components/FriendSelector/styles.ts
