+++ src/pages/Home/components/RankingSection/styles.ts
